Handle more concurrent events on windows.
Do not arbitrary limit the number of even the windows layer can
handle.  We make use of a vector to grow the number of events
with the number of connection.

Unfortunately, std::vector.data() and
std::vector::resize(size, value) are C++ features, we will need to force
C++11 on the whole project.